Functionality for Randomized passwords (test before merge) (#71)
* Created a randomize password function

created a randomized string function and put it in data/storageExampleItems.kt. it gets displayed in the storage view

* Added a button for random password

This button is only added to the "add credentials page" (storage_item_popup_add). It has not been tested yet.

* Moved code to correct popup view

* Added the randomize password button to both the edit and add credential page

Added the randomize password button to both the add credentials and edit credentials views.

* Update gradle.xml

* Update misc.xml

---------

Co-authored-by: JoXW100 <joxw100@gmail.com>